Use String over CharSequence
Interacting with CharSequence can result in some syntactical overhead compared to interacting with Strings.
As an example, when using the tagName from the githubRelease extension, passing that value to another API requires converting that value to a String.
This tends to add some amount of noise to the code as most APIs use Strings for such things.